+++ source/Internal/Domain/Authentication/Bridge/PasswordServiceBridgeInterface.php
@@ -1,0 +1,38 @@
+<?php declare(strict_types=1);
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Authentication\Bridge;
+
+/**
+ * @internal
+ */
+interface PasswordServiceBridgeInterface
+{
+    /**
+     * @param string $password
+     *
+     * @return string
+     */
+    public function hash(string $password): string;
+
+
+    /**
+     * @param string $passwordHash
+     *
+     * @return bool
+     */
+    public function passwordNeedsRehash(string $passwordHash): bool;
+
+    /**
+     * Verify that a given password matches a given hash
+     *
+     * @param string $password
+     * @param string $passwordHash
+     *
+     * @return bool
+     */
+    public function verifyPassword(string $password, string $passwordHash): bool;
+}