+++ source/Internal/Domain/Authentication/Exception/UnavailablePasswordHashException.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Authentication\Exception;
+
+/**
+ * Class UnavailablePasswordHashException
+ */
+class UnavailablePasswordHashException extends \Exception
+{
+}