+++ source/Internal/Domain/Authentication/Service/PasswordHashServiceInterface.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Authentication\Service;
+
+/**
+ * @internal
+ */
+interface PasswordHashServiceInterface
+{
+    /**
+     * @param string $password
+     *
+     * @return string
+     */
+    public function hash(string $password): string;
+
+    /**
+     * @param string $passwordHash
+     *
+     * @return bool
+     */
+    public function passwordNeedsRehash(string $passwordHash): bool;
+}