+++ source/Internal/Domain/Authentication/Service/PasswordVerificationServiceInterface.php
@@ -1,0 +1,25 @@
+<?php declare(strict_types=1);
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Domain\Authentication\Service;
+
+/**
+ * Class PasswordVerificationService
+ *
+ * @internal
+ */
+interface PasswordVerificationServiceInterface
+{
+    /**
+     * Verify that a given password matches a given hash
+     *
+     * @param string $password
+     * @param string $passwordHash
+     *
+     * @return bool
+     */
+    public function verifyPassword(string $password, string $passwordHash): bool;
+}