+++ tests/Unit/Internal/Form/ContactForm/ContactFormEmailValidatorTest.php
@@ -4,11 +4,11 @@
  * See LICENSE file for license details.
  */
 
-namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Form\ContactForm;
+namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Domain\Contact\Form;
 
 use OxidEsales\EshopCommunity\Internal\Common\Form\Form;
 use OxidEsales\EshopCommunity\Internal\Common\Form\FormField;
-use OxidEsales\EshopCommunity\Internal\Form\ContactForm\ContactFormEmailValidator;
+use OxidEsales\EshopCommunity\Internal\Domain\Contact\Form\ContactFormEmailValidator;
 use OxidEsales\EshopCommunity\Internal\Adapter\ShopAdapter;
 
 class ContactFormEmailValidatorTest extends \PHPUnit\Framework\TestCase