+++ tests/Unit/Internal/Form/ContactForm/ContactFormFactoryTest.php
@@ -4,7 +4,7 @@
  * See LICENSE file for license details.
  */
 
-namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Form\ContactForm;
+namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Domain\Contact\Form;
 
 use OxidEsales\EshopCommunity\Internal\Adapter\ShopAdapterInterface;
 use OxidEsales\EshopCommunity\Internal\Common\Form\FormField;
@@ -12,8 +12,8 @@ use OxidEsales\EshopCommunity\Internal\Common\Form\RequiredFieldsValidator;
 use OxidEsales\EshopCommunity\Internal\Common\FormConfiguration\FieldConfiguration;
 use OxidEsales\EshopCommunity\Internal\Common\FormConfiguration\FormConfiguration;
 use OxidEsales\EshopCommunity\Internal\Common\FormConfiguration\FormConfigurationInterface;
-use OxidEsales\EshopCommunity\Internal\Form\ContactForm\ContactFormEmailValidator;
-use OxidEsales\EshopCommunity\Internal\Form\ContactForm\ContactFormFactory;
+use OxidEsales\EshopCommunity\Internal\Domain\Contact\Form\ContactFormEmailValidator;
+use OxidEsales\EshopCommunity\Internal\Domain\Contact\Form\ContactFormFactory;
 use OxidEsales\EshopCommunity\Internal\Common\Form\FormInterface;
 
 class ContactFormFactoryTest extends \PHPUnit\Framework\TestCase